requirements.txt
flask
reportlab
pyyaml
pillow
pywebpush
apscheduler
cryptography
beautifulsoup4
